pyproject.toml
[project]
name = "echo_chamber"
version = "0.1.0"
description = "Code and examples for Echo Chamber LLM Jailbreak"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pydantic>=2.11.7",
    "tqdm>=4.67.1",
    "uv>=0.6.1",
]

[project.optional-dependencies]
google = [
    "google-genai>=1.28.0",
]
ollama = [
    "ollama>=0.5.1",
    "httpx>=0.28.1",
]
openai = [
    "openai>=1.98.0",
]
dev = [
    "pre-commit>=4.2.0",
    "pyrefly>=0.27.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.7",
]


[tool.pyrefly]
project-includes = ["echo_chamber/**/*"]
project-excludes = ["**/*venv/**/*"]

[tool.ruff]
extend-include = ["*.ipynb"]
line-length = 85

[tool.ruff.lint]
extend-select = [
  "I", # isort
  "F", # Pyflakes
  "D", # Pydocstyle
]
ignore = ["D100", "D104", "D401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "D104"]
"**/{tests,docs,tools}/*" = ["E402", "D"]


[tool.setuptools.packages.find]
where = ["."]
include = ["echo_chamber*"]
